requirement.txt
streamlit
python-chess
numpy==1.24.0
scikit-learn==1.1.3
xgboost==1.4.2
shap==0.39.0
pandas==1.5.3
matplotlib==3.6.3